Turned off auto-disconnect, fixed client status bugs

+++ src/com/scs/stetech1/netmessages/GameSuccessfullyJoinedMessage.java
@@ -5,7 +5,7 @@ import com.jme3.network.serializing.Serializable;
 @Serializable
 public class GameSuccessfullyJoinedMessage extends MyAbstractMessage {
 
-	public long avatarEntityID;
+	public int avatarEntityID;
 	public int playerID;
 
 	public GameSuccessfullyJoinedMessage() {
@@ -13,7 +13,7 @@ public class GameSuccessfullyJoinedMessage extends MyAbstractMessage {
 	}
 	
 	
-	public GameSuccessfullyJoinedMessage(int _playerID, long _avatarEntityID) {
+	public GameSuccessfullyJoinedMessage(int _playerID, int _avatarEntityID) {
 		super(true);
 		
 		playerID = _playerID;
